requirements.txt
torch
python-dotenv